+++ packages/openapi-ts/src/compiler/index.ts
@@ -1,7 +1,7 @@
 import { type PathLike, rmSync, writeFileSync } from 'node:fs';
 import path from 'node:path';
 
-import ts from 'typescript';
+import type ts from 'typescript';
 
 import { ensureDirSync } from '../generate/utils';
 import * as classes from './classes';
@@ -178,7 +178,6 @@ export const compiler = {
   conditionalExpression: types.createConditionalExpression,
   constVariable: module.createConstVariable,
   constructorDeclaration: classes.createConstructorDeclaration,
-  elementAccessExpression: transform.createElementAccessExpression,
   enumDeclaration: types.createEnumDeclaration,
   exportAllDeclaration: module.createExportAllDeclaration,
   exportNamedDeclaration: module.createNamedExportDeclarations,