+++ packages/openapi-ts/src/generate/client.ts
@@ -1,22 +1,31 @@
-import { copyFileSync, readFileSync, writeFileSync } from 'node:fs';
-import { createRequire } from 'node:module';
+import { copyFileSync } from 'node:fs';
 import path from 'node:path';
 
-import { getConfig, isStandaloneClient } from '../utils/config';
-import { ensureDirSync } from './utils';
+import { getConfig, isLegacyClient } from '../utils/config';
+import { ensureDirSync, relativeModulePath } from './utils';
 
-const require = createRequire(import.meta.url);
-
-export const clientModulePath = () => {
+export const clientModulePath = ({
+  sourceOutput,
+}: {
+  sourceOutput: string;
+}) => {
   const config = getConfig();
-  return config.client.bundle ? './client' : config.client.name;
+
+  if (config.client.bundle) {
+    return relativeModulePath({
+      moduleOutput: 'client',
+      sourceOutput,
+    });
+  }
+
+  return config.client.name;
 };
 
 export const clientOptionsTypeName = () => 'Options';
 
 /**
  * (optional) Creates a `client.ts` file containing the same exports as a
- * standalone client package. Creates a `core` directory containing the modules
+ * standalone client package. Creates a `client` directory containing the modules
  * from standalone client. These files are generated only when `client.bundle`
  * is set to true.
  */
@@ -26,12 +35,12 @@ export const generateClient = async (
 ) => {
   const config = getConfig();
 
-  if (!isStandaloneClient(config) || !config.client.bundle) {
+  if (isLegacyClient(config) || !config.client.bundle) {
     return;
   }
 
   // create directory for client modules
-  const dirPath = path.resolve(outputPath, 'core');
+  const dirPath = path.resolve(outputPath, 'client');
   ensureDirSync(dirPath);
 
   const clientModulePath = path.normalize(require.resolve(moduleName));
@@ -52,12 +61,4 @@ export const generateClient = async (
       path.resolve(dirPath, file),
     );
   });
-
-  // copy index file with cherry-picked exports
-  const nodeIndexFile = readFileSync(
-    path.resolve(clientSrcPath, 'node', 'index.ts'),
-    'utf-8',
-  );
-  const indexFile = nodeIndexFile.replaceAll('../', './core/');
-  writeFileSync(path.resolve(outputPath, 'client.ts'), indexFile, 'utf-8');
 };