+++ packages/openapi-ts/src/node/index.ts
@@ -1,2 +1,0 @@
-export { createClient, defineConfig } from '../';
-export type { UserConfig } from '../types/config';