+++ packages/openapi-ts/src/openApi/3.1/index.ts
@@ -1,0 +1,2 @@
+export { parseV3_1 } from './parser';
+export type { OpenApiV3_1 } from './types/spec';