+++ packages/openapi-ts/src/utils/parse.ts
@@ -1,16 +1,14 @@
 import type { Operation, OperationParameter } from '../openApi';
 import { sanitizeNamespaceIdentifier } from '../openApi';
 import { camelCase } from './camelCase';
-import { getConfig, isStandaloneClient } from './config';
+import { getConfig, isLegacyClient } from './config';
 import { transformTypeKeyName } from './type';
 
-export const operationFilterFn = (operation: Operation): boolean => {
+export const operationFilterFn = (operationKey: string): boolean => {
   const config = getConfig();
-
   const regexp = config.services.filter
     ? new RegExp(config.services.filter)
     : undefined;
-  const operationKey = `${operation.method} ${operation.path}`;
   return !regexp || regexp.test(operationKey);
 };
 
@@ -21,7 +19,7 @@ export const operationParameterFilterFn = (
 
   // legacy clients ignore the "api-version" param since we do not want to
   // add it as the first/default parameter for each of the service calls
-  return isStandaloneClient(config) || parameter.prop !== 'api-version';
+  return !isLegacyClient(config) || parameter.prop !== 'api-version';
 };
 
 /**
@@ -42,7 +40,7 @@ export const operationNameFn = (operation: Omit<Operation, 'name'>): string => {
 
   // legacy clients ignore the "api-version" param since we do not want to
   // add it as the first/default parameter for each of the service calls
-  if (!isStandaloneClient(config)) {
+  if (isLegacyClient(config)) {
     urlWithoutPlaceholders = urlWithoutPlaceholders.replace(
       /[^/]*?{api-version}.*?\//g,
       '',
@@ -52,7 +50,7 @@ export const operationNameFn = (operation: Omit<Operation, 'name'>): string => {
   urlWithoutPlaceholders = urlWithoutPlaceholders
     .replace(/{(.*?)}/g, 'by-$1')
     // replace slashes with hyphens for camelcase method at the end
-    .replace(/\//g, '-');
+    .replace(/[/:]/g, '-');
 
   return camelCase({
     input: `${operation.method}-${urlWithoutPlaceholders}`,
@@ -64,7 +62,7 @@ export const operationParameterNameFn = (
 ): string => {
   const config = getConfig();
 
-  return isStandaloneClient(config)
+  return !isLegacyClient(config)
     ? parameter.prop
     : transformTypeKeyName(parameter.prop);
 };