+++ packages/openapi-ts/test/sample.cjs
@@ -1,7 +1,7 @@
 const path = require('node:path');
 
 const main = async () => {
-  /** @type {import('../src/node/index').UserConfig} */
+  /** @type {import('../src').UserConfig} */
   const config = {
     client: {
       // bundle: true,
@@ -9,9 +9,10 @@ const main = async () => {
       name: '@hey-api/client-fetch',
     },
     // debug: true,
-    experimental_parser: true,
+    // experimental_parser: true,
     // input: './test/spec/v3-transforms.json',
     input: './test/spec/v3.json',
+    // input: './test/spec/3.1.0/required-all-of-ref.json',
     // input: './test/spec/v2.json',
     // input: 'https://mongodb-mms-prod-build-server.s3.amazonaws.com/openapi/2caffd88277a4e27c95dcefc7e3b6a63a3b03297-v2-2023-11-15.json',
     // name: 'foo',
@@ -19,12 +20,9 @@ const main = async () => {
       path: './test/generated/sample/',
     },
     plugins: [
-      {
-        // infiniteQueryOptions: false,
-        // mutationOptions: false,
-        name: '@tanstack/react-query',
-        // queryOptions: false,
-      },
+      // '@tanstack/react-query',
+      // '@hey-api/services',
+      'zod',
     ],
     schemas: {
       export: false,
@@ -32,7 +30,7 @@ const main = async () => {
     services: {
       // export: false,
       // asClass: true,
-      // filter: '^POST /api/v{api-version}/upload$',
+      // filter: '^GET /api/v{api-version}/simple:operation$',
       // export: false,
       // name: '^Parameters',
     },
@@ -43,13 +41,13 @@ const main = async () => {
       // include:
       //   '^(_400|CompositionWithOneOfAndProperties)',
       // name: 'PascalCase',
-      tree: false,
+      // tree: false,
     },
     // useOptions: false,
   };
 
   const { createClient } = await import(
-    path.resolve(process.cwd(), 'dist/node/index.cjs')
+    path.resolve(process.cwd(), 'dist', 'index.cjs')
   );
   await createClient(config);
 };